Add embedded Postgres for testing
We can successfully import schema and data but the current solution
is not optimal. I had to delete all "alter table" stables from
schema.sql that included the user 'developer'. Need to find a way
to dump the whole real database along with the user and import it
into the embedded database.