Basic queries (#3)
* Add model peristence classes

Continue with building the Dao classes and make some unit tests.
Try adding and selecting users first.

* Refactor schema

Use bigints for primary keys. Found out how to properly export and
import schema without any erros. Instructions for this added to
README.

* Refactor DTOs

Changed IDs to long

* Add UserDao.save()

Refactored table names into plural.

* Add UserDAO get methods for id and email

* Add docker postgres

Instructions added to README

* Add UserDao.getAll()

* Add UserDao.update()

* Add UserDao.delete()

* Start BookDao

Create getAll()

* Add BookDao get()

* Finish BookDao

* Add ChapterDAO.get()

* Add insertChapters()

Use this to add all rows needed to chapters table.

* Finish insertChapters()

Need to run this on live DB to build chapters.

* Add all rows to chapters table

* Finish ChapterDao

* Add NoteDao.save()

* Add update/get methods to NoteDao

* Refactor UserDao

* Fix primary key bug on users in embedded db

No way to export in datagrip the schema with an option to exclude
explicit values for primary keys. As a result, the nextval() is not
called on their values which causes an error when trying to import.

* Add NoteDao.delete()

* Add ViewDao get and save methods

Had to add uuid-ossp extension to DB to generate UUIDs for views.

* Add ViewDao.getUnique()

* Add ViewDao.getAll()

Should be finished with this class.

* Refactor UpdatableDao.save()

Now returns id of the relation.

* Save progress

Create DaoService

Refactor Dao.getUnique to return List. Need to update all
implementing classes.

* Refactor all Dao.get(Map<String, Object>) methods

* Refactor unit tests

* Add CommentDao.update()

* Finish CommentDao

* Add addUserNotesToView

InDaoServiceJbcImpl class. Need to test this.

* Save progress

In the middle of debugging addUserNotesToView. Try to make the unit
test pass.

* Finish DaoService.addUserNotesToView()