Basic security (#8)
* Set up in mem authentication

Use this curl command to working test:

curl -u admin
http://localhost:8080/biblestudy/6e9e6366-f386-11e9-b633-0242ac110002/Genesis/1

* Enable CORS on security config

This allows frontend to work

* Add DB based authentication without enconding

Currently using noop for encoding. Need to try with Bcrypt encoder
and fix some of the columns used in the sql in the security config.

* Refactor js scripts into main.js

* Get user principals by email

Currently using session cookie for requestions without auth
header. Looks like csrf may not be implement correctly by Spring
for Rest Services. Look at this article:
https://www.codesandnotes.be/2015/02/05/spring-securitys-csrf-protection-for-rest-services-the-client-side-and-the-server-side/

* Configure CORS only for localhost

* Add serve XSRF token to response

Need to add it as a param in ajax query

* Add working post request to frontend

CSRF disabled :(

* Remove auth header for post request

Session needs to be established first before post can work. This is
good except that it is CSRF vulnerable.

* Add BCrypt encoding for pass hash

* Add working CSRF protection

X-XSRF-TOKEN is the header needed, NOT X-CSRF-TOKEN...