Views frontend (#20)
* Add integration test suite

* Add getViewCodes to controller

Todo- hook this up to the frontend when the user clicks on manage
views.

* Add get user views to dashboard

* Add get notes to manage view dashboard

* Add remove note from view on frontend

Not connected to backend yet.

* Finish remove note from view in dashboard

Continue with creating and deleting entire views.

* Add create new view to dashboard

* Add delete current view to dashboard

* Start add notes section on frontend

* Save progress

Continue with showAddNotes() in main.js.

* Add select authors and set minimum ranking

Only cosmetic on dashboard. Need to connect to the endpoints.

* Save progress

Added addNotesToView() service and controller methods. Need to
implement ajax request.

* Finish add notes to current view

Works frontend backend.