Jwt (#34)
* Add auth server

At this point we can hit up the auth server for a token and verify it

* Make persistenceservice resource server

* Add public urls to resource server

* Wrestle with Feign

Can propogate token header. Looks like OAuth2ClientContext may not
be instantiated correctly

* Propagate bearer token through client

Mybiblestudyservice will send the token through Zuul to the
protected persistenceservice and validate it against the auth
server.

TODO: implement login functionality and switch to JWT

* Add JWT to auth server

* Add login/logout functionality for jwt

TODO: Mybiblestudyservice acts as the web client that authenticates
against the auth server for the JWT. It does it directly with
RestTemplate. Might want to change it to go through Zuul instead
for proper load balancing, short circuiting and bulkhead strategies.

* Move some properties to conf server