Dockercompose (#37)
* Start docker-compose

TODO: Add env variables for conf to all bootstrap files and eureka,
zookeeper and kafka variables to application properities

* Update boostrap and application files

TODO: Remount the local postgres volume with docker-compose

Check and see if uri for the auth_server needs to be updated for
docker.

* Fix authserver db connection

* Add run.sh to authserver dockerfile

Need to add a ping and wait against the config server to make sure
auth server isn't started before the config server is ready.

* Save progress

bibletextservice cannot connect to its db with docker compose.

* Fix reading of Bible csv files

Could not load them as classpath resources in the BibleTextService
docker container. Manually copied them over in the docker file and
loaded them with an absolute path.

* Fix prim key gen strategy for bibletextservice

* Fix postgres-bibles port in dockercompose file

* Fix persistence service w dockercompose

* Save progress

Continue with debugging mybiblestudy service. Trying running
everything with dockercompose first, then try debugging.

* Fix kafka server location for mybiblestudyservice

Wrong property was set in application.yml

* Fix zuul config

Now the whole thing seems to be working with docker-compose.

* Fix bibletextservice for docker and non docker

Should now run in both environments. Took out the building of the
bibltext db tables upon start up and mounted it as a volume on the
local file system.