Sleuth (#39)
* Add Sleuth dependencies

ISSUE: Overwrites my custom handling of the jwt in feign. Try
using a Ribbon backed RestTemplate instead of feign.

* Fix Sleuth

Sleuth now works. For some reason user context stopped getting
passed to the Hystrix thread pool used by Feign so we couldn't
append the JWT to the header. This was solved by creating a session
bean for the JWT and passing it into the call to the Feign client.