Merge Microservices (#41)
* Add parent pom

* Add confsvr module

* Fix cloud config dependency problem

* Fix intellij mvn dependency problems

* Add refreshscope to service

MyBibleStudyWebAppService is currently booting. Need to configure
confsvr.

* Add working config for local repo

* Add remote repo for cloud config

* Add encryption to confsvr

* Add Eureka

* Refactor application.properties to .yml

* Add feignclient

* Add BibleTextClient

This implements the FeignClient and calls the bibletextservice
to get Bible verses. Currently working in dev environment.

* Add Hystrix

Implements:

Circuitbreaker
Fallback
Bulkhead

* Add UserContext

In the utils package with other supporting classes

* Add context injection to Hystrix threads

* Add correlationId to UserContext

* Add Zuul

* Fix Zuul routes

* Add utils module

Build needs to occur from parent directory

* Fix script path for unit tests

* Add Zuul filters

* Add authenticationservice

* Add authservice issues token on good password

* Add BCrypt to authservice

* Persistenceservice (#33)

* Add persistenceservice

* Refactor DaoService

Remove async methods

* Add PersistenceController

Refactored DaoService removing exceptions

* Add PersistenceClient

* Add PersistenceService

* Refactor MainServiceImpl

* Fix connection to PersistenceService

TODO: Look at login functionality and debug

* Fix login functionality

* Fix PersistenceController

Note to self- Don't use objects as path variables

* Remove unused feign config files

Also cleaned up some unused POM dependencies

* Jwt (#34)

* Add auth server

At this point we can hit up the auth server for a token and verify it

* Make persistenceservice resource server

* Add public urls to resource server

* Wrestle with Feign

Can propogate token header. Looks like OAuth2ClientContext may not
be instantiated correctly

* Propagate bearer token through client

Mybiblestudyservice will send the token through Zuul to the
protected persistenceservice and validate it against the auth
server.

TODO: implement login functionality and switch to JWT

* Add JWT to auth server

* Add login/logout functionality for jwt

TODO: Mybiblestudyservice acts as the web client that authenticates
against the auth server for the JWT. It does it directly with
RestTemplate. Might want to change it to go through Zuul instead
for proper load balancing, short circuiting and bulkhead strategies.

* Move some properties to conf server

* Redis (#35)

* Add JPA classes and bible text into H2

This is for the bibletextservice

* Load schema and data

TODO
-Check fetch strategy
-Fix text (verses get chopped off due to commas and other chars)
-Create a text table that maps to verse instead

* Fix verse input parse a bit

Gets the whole line but there is a problem with the csv file that
puts in a lot of extra quotation marks for some reason.

* Add VerseText

This is a new table to contain texts from difference translations
to match to each verse.

Also refactor the db initialization to a config class.

* Add docker run script and schema diagram

* Refactor Version entity to Translation

* Start redis

Getting error about not getting resource from the pool

* Add Redis docker to run script

* Use redis:alpine

* Add CacheService

* Add Redis to Persistenceservice

Probably not too useful and can be removed

* Fix persistenceservice unittests

* Kafka (#36)

* Add kafka to docker run script

* Add new kafka docker run command

* Use docker-compose for Kafka setup

* Fix Kafka subscriber

The handler cannot be in the main class

* Dockercompose (#37)

* Start docker-compose

TODO: Add env variables for conf to all bootstrap files and eureka,
zookeeper and kafka variables to application properities

* Update boostrap and application files

TODO: Remount the local postgres volume with docker-compose

Check and see if uri for the auth_server needs to be updated for
docker.

* Fix authserver db connection

* Add run.sh to authserver dockerfile

Need to add a ping and wait against the config server to make sure
auth server isn't started before the config server is ready.

* Save progress

bibletextservice cannot connect to its db with docker compose.

* Fix reading of Bible csv files

Could not load them as classpath resources in the BibleTextService
docker container. Manually copied them over in the docker file and
loaded them with an absolute path.

* Fix prim key gen strategy for bibletextservice

* Fix postgres-bibles port in dockercompose file

* Fix persistence service w dockercompose

* Save progress

Continue with debugging mybiblestudy service. Trying running
everything with dockercompose first, then try debugging.

* Fix kafka server location for mybiblestudyservice

Wrong property was set in application.yml

* Fix zuul config

Now the whole thing seems to be working with docker-compose.

* Fix bibletextservice for docker and non docker

Should now run in both environments. Took out the building of the
bibltext db tables upon start up and mounted it as a volume on the
local file system.

* Add integration test for mybiblestudy service

* Refactor config files (#38)

* Add invalidate cache on note update

* Sleuth (#39)

* Add Sleuth dependencies

ISSUE: Overwrites my custom handling of the jwt in feign. Try
using a Ribbon backed RestTemplate instead of feign.

* Fix Sleuth

Sleuth now works. For some reason user context stopped getting
passed to the Hystrix thread pool used by Feign so we couldn't
append the JWT to the header. This was solved by creating a session
bean for the JWT and passing it into the call to the Feign client.

* Elk (#40)

* Add working Elastic stack for logging

* Add logback encoder to all services

TODO: Change the mounted data volumes for Elasticsearch and
Filebeat in the docker-compose file to outside the project directory.

* Refactor elastic stack directory

Mount volumes to /var/lib/* instead of inside project directory.

* Update documentation